# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/menu-categories": {
      "get": {
        "description": "Возвращает список всех категорий меню, отсортированных по названию",
        "operationId": "MenuCategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список категорий успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuCategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Получить все категории меню",
        "tags": [
          "Menu Categories"
        ]
      },
      "post": {
        "description": "Создает новую категорию меню",
        "operationId": "MenuCategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Категория успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuCategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные для создания категории"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать новую категорию",
        "tags": [
          "Menu Categories"
        ]
      }
    },
    "/menu-categories/{id}": {
      "get": {
        "description": "Возвращает детальную информацию о категории меню",
        "operationId": "MenuCategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор категории",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Категория найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuCategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Категория с указанным ID не найдена"
          }
        },
        "summary": "Получить категорию по ID",
        "tags": [
          "Menu Categories"
        ]
      },
      "patch": {
        "description": "Обновляет информацию о категории меню",
        "operationId": "MenuCategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор категории",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Категория успешно обновлена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuCategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные для обновления категории"
          },
          "404": {
            "description": "Категория с указанным ID не найдена"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить категорию",
        "tags": [
          "Menu Categories"
        ]
      },
      "delete": {
        "description": "Удаляет категорию меню",
        "operationId": "MenuCategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор категории",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Категория успешно удалена"
          },
          "404": {
            "description": "Категория с указанным ID не найдена"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить категорию",
        "tags": [
          "Menu Categories"
        ]
      }
    },
...
    "/menu/{id}": {
      "get": {
        "description": "Возвращает детальную информацию о блюде с информацией о категории",
        "operationId": "MenuController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор блюда",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Блюдо найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Блюдо с указанным ID не найдено"
          }
        },
        "summary": "Получить блюдо по ID",
        "tags": [
          "Menu"
        ]
      },
      "patch": {
        "description": "Обновляет информацию о блюде",
        "operationId": "MenuController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор блюда",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Блюдо успешно обновлено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные для обновления блюда"
          },
          "404": {
            "description": "Блюдо с указанным ID не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить блюдо",
        "tags": [
          "Menu"
        ]
      },
      "delete": {
        "description": "Полностью удаляет блюдо из базы данных",
        "operationId": "MenuController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор блюда",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Блюдо успешно удалено"
          },
          "404": {
            "description": "Блюдо с указанным ID не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить блюдо",
        "tags": [
          "Menu"
        ]
      }
    },
...
  },
  "info": {
    "title": "Georgian Restaurant API",
    "description": "API для управления грузинскими ресторанами с системой бронирования, заказов и аудита",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "MenuCategoryResponseDto": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "number",
            "description": "Уникальный идентификатор категории",
            "example": 1
          },
          "category_name": {
            "type": "string",
            "description": "Название категории меню",
            "example": "Горячие блюда",
            "minLength": 2,
            "maxLength": 100
          },
          "category_description": {
            "type": "string",
            "description": "Описание категории",
            "example": "Основные блюда грузинской кухни"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания категории",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата последнего обновления",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "category_id",
          "category_name",
          "created_at",
          "updated_at"
        ]
      },
      "CreateMenuCategoryDto": {
        "type": "object",
        "properties": {
          "category_name": {
            "type": "string",
            "description": "Название категории меню",
            "example": "Горячие блюда",
            "minLength": 2,
            "maxLength": 100
          },
          "category_description": {
            "type": "string",
            "description": "Описание категории",
            "example": "Основные блюда грузинской кухни"
          }
        },
        "required": [
          "category_name"
        ]
      },
      "UpdateMenuCategoryDto": {
        "type": "object",
        "properties": {
          "category_name": {
            "type": "string",
            "description": "Название категории меню",
            "example": "Горячие блюда",
            "minLength": 2,
            "maxLength": 100
          },
          "category_description": {
            "type": "string",
            "description": "Описание категории",
            "example": "Основные блюда грузинской кухни"
          }
        }
      },
      "MenuItemResponseDto": {
        "type": "object",
        "properties": {
          "item_id": {
            "type": "number",
            "description": "Уникальный идентификатор блюда",
            "example": 1
          },
          "item_name": {
            "type": "string",
            "description": "Название блюда",
            "example": "Хачапури по-аджарски",
            "minLength": 3,
            "maxLength": 150
          },
          "item_description": {
            "type": "string",
            "description": "Описание блюда",
            "example": "Традиционная грузинская лепешка с сыром и яйцом"
          },
          "category_id": {
            "type": "number",
            "description": "ID категории меню",
            "example": 1
          },
          "price": {
            "type": "number",
            "description": "Стоимость блюда",
            "example": 25.5
          },
          "cooking_time_minutes": {
            "type": "number",
            "description": "Время приготовления в минутах",
            "example": 15
          },
          "calories": {
            "type": "number",
            "description": "Калории",
            "example": 350
          },
          "is_vegetarian": {
            "type": "boolean",
            "description": "Вегетарианское блюдо",
            "example": false
          },
          "is_spicy": {
            "type": "boolean",
            "description": "Острое блюдо",
            "example": false
          },
          "is_deleted": {
            "type": "boolean",
            "description": "Удалено (мягкое удаление)",
            "example": false
          },
          "image_url": {
            "type": "string",
            "description": "URL изображения блюда",
            "example": "https://example.com/image.jpg"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата последнего обновления",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "category": {
            "type": "object",
            "description": "Информация о категории"
          }
        },
        "required": [
          "item_id",
          "item_name",
          "category_id",
          "price",
          "cooking_time_minutes",
          "is_vegetarian",
          "is_spicy",
          "is_deleted",
          "created_at",
          "updated_at"
        ]
      },
...
      "UpdateMenuItemDto": {
        "type": "object",
        "properties": {
          "item_name": {
            "type": "string",
            "description": "Название блюда",
            "example": "Хачапури по-аджарски",
            "minLength": 3,
            "maxLength": 150
          },
          "item_description": {
            "type": "string",
            "description": "Описание блюда",
            "example": "Традиционная грузинская лепешка с сыром и яйцом"
          },
          "category_id": {
            "type": "number",
            "description": "ID категории меню",
            "example": 1
          },
          "price": {
            "type": "number",
            "description": "Стоимость блюда",
            "example": 25.5
          },
          "cooking_time_minutes": {
            "type": "number",
            "description": "Время приготовления в минутах",
            "example": 15
          },
          "calories": {
            "type": "number",
            "description": "Калории",
            "example": 350
          },
          "is_vegetarian": {
            "type": "boolean",
            "description": "Вегетарианское блюдо",
            "example": false
          },
          "is_spicy": {
            "type": "boolean",
            "description": "Острое блюдо",
            "example": false
          },
          "image_url": {
            "type": "string",
            "description": "URL изображения блюда",
            "example": "https://example.com/image.jpg"
          }
        }
      },
...
    }
  }
}